deleted the denifitions folder. about time. created wait_for_ajax.rb but it's not working without a driver for capybara. tried selenium, capybara-webkit is archived, trying apparition next. my p key is breaking. again. pppppppppppppppppppppppppp
