+++ features/step_definitions/tutee_steps.rb
@@ -16,44 +16,6 @@ Given /the following tutees exist/ do |tutees_table|
   end
 end
 
-Given /^(?:|I )am on (.+)$/ do |page_name|
-  visit path_to(page_name)
-end
-
-And /^(?:|I )fill in "([^"]*)" with "([^"]*)"$/ do |field, value|
-  fill_in(field, :with => value)
-end
-
-And /^(?:|I )change "([^"]*)" to "([^"]*)"$/ do |field, value|
-  fill_in(field, :with => value)
-end
-
-Then(/^I should see "(.*?)"$/) do |arg1|
-  page.should have_content("#{arg1}")
-end
-
-When /^(?:|I )press "([^"]*)"$/ do |button|
-  click_button(button)
-end
-
-When /^(?:|I )press link "([^"]*)"$/ do |link|
-  click_link(link)
-end
-
-And /^(?:|I )should be on (.+)$/ do |page_name|
-  current_path = URI.parse(current_url).path
-  if current_path.respond_to? :should
-    current_path.should == path_to(page_name)
-  else
-    assert_equal path_to(page_name), current_path
-  end
-end
-
-And /^(?:|I )select "([^"]*)" from "([^"]*)"$/ do |value, field|
-  select(value, :from => field)
-end
-
-
 Given /I login as "(.*)"/ do |name|
   step %{I am on the login page}
   step %{I fill in "username" with "#{Tutee.find_by_first_name(name).email}"}
