--- features/step_denifition/web_steps.rb
@@ -1,3 +1,0 @@
-# Given /^(?:|I )am on the user page for "(.*)"/ do |email|
-#   visit tutee_path(Tutee.find_by_email(email))
-# end
